refactor: add ActionInitializerItem to replace ActionInitializer
