+++ packages/core/client/src/modules/actions/add-child/CreateChildInitializer.tsx
@@ -8,7 +8,7 @@
  */
 
 import React from 'react';
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 export const CreateChildInitializer = (props) => {
   const schema = {
     type: 'void',
@@ -64,5 +64,5 @@ export const CreateChildInitializer = (props) => {
       },
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
