+++ packages/core/client/src/modules/actions/expand-collapse/ExpandableActionInitializer.tsx
@@ -8,7 +8,7 @@
  */
 
 import React from 'react';
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 
 export const ExpandableActionInitializer = (props) => {
   const schema = {
@@ -29,5 +29,5 @@ export const ExpandableActionInitializer = (props) => {
       },
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
