+++ packages/core/client/src/modules/actions/filter/FilterActionInitializer.tsx
@@ -8,8 +8,7 @@
  */
 
 import React from 'react';
-
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 
 export const FilterActionInitializer = (props) => {
   const schema = {
@@ -24,5 +23,5 @@ export const FilterActionInitializer = (props) => {
       icon: 'FilterOutlined',
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
