+++ packages/core/client/src/modules/actions/submit/CreateSubmitActionInitializer.tsx
@@ -8,7 +8,7 @@
  */
 
 import React from 'react';
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 
 export const CreateSubmitActionInitializer = (props) => {
   const schema = {
@@ -26,5 +26,5 @@ export const CreateSubmitActionInitializer = (props) => {
       triggerWorkflows: [],
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
