+++ packages/core/client/src/modules/actions/submit/UpdateSubmitActionInitializer.tsx
@@ -8,8 +8,7 @@
  */
 
 import React from 'react';
-
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 
 export const UpdateSubmitActionInitializer = (props) => {
   const schema = {
@@ -28,5 +27,5 @@ export const UpdateSubmitActionInitializer = (props) => {
       triggerWorkflows: [],
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
