+++ packages/core/client/src/modules/actions/view-edit-popup/UpdateActionInitializer.tsx
@@ -8,8 +8,7 @@
  */
 
 import React from 'react';
-
-import { ActionInitializer } from '../../../schema-initializer/items/ActionInitializer';
+import { ActionInitializerItem } from '../../../schema-initializer/items/ActionInitializerItem';
 
 export const UpdateActionInitializer = (props) => {
   const schema = {
@@ -59,5 +58,5 @@ export const UpdateActionInitializer = (props) => {
       },
     },
   };
-  return <ActionInitializer {...props} schema={schema} />;
+  return <ActionInitializerItem {...props} schema={schema} />;
 };
