+++ packages/core/client/src/schema-initializer/items/ActionInitializer.tsx
@@ -7,23 +7,18 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { merge } from '@formily/shared';
 import React from 'react';
 
-import { SchemaInitializerItem, useSchemaInitializer, useSchemaInitializerItem } from '../../application';
+import { useSchemaInitializerItem } from '../../application';
+import { InitializerWithSwitch } from './InitializerWithSwitch';
 
+/**
+ * @deprecated
+ * use ActionInitializerItem instead
+ * @param props
+ * @returns
+ */
 export const ActionInitializer = (props) => {
   const itemConfig = useSchemaInitializerItem();
-  const { insert } = useSchemaInitializer();
-
-  return (
-    <SchemaInitializerItem
-      title={itemConfig.title}
-      onClick={() => {
-        const s = merge(props.schema || {}, itemConfig.schema || {});
-        itemConfig?.schemaInitialize?.(s);
-        insert(s);
-      }}
-    />
-  );
+  return <InitializerWithSwitch {...itemConfig} {...props} item={itemConfig} type={'x-action'} />;
 };
