+++ packages/core/client/src/schema-initializer/items/ActionInitializerItem.tsx
@@ -1,0 +1,29 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+import { merge } from '@formily/shared';
+import React from 'react';
+
+import { SchemaInitializerItem, useSchemaInitializer, useSchemaInitializerItem } from '../../application';
+
+export const ActionInitializerItem = (props) => {
+  const itemConfig = useSchemaInitializerItem();
+  const { insert } = useSchemaInitializer();
+
+  return (
+    <SchemaInitializerItem
+      title={itemConfig.title}
+      onClick={() => {
+        const s = merge(props.schema || {}, itemConfig.schema || {});
+        itemConfig?.schemaInitialize?.(s);
+        insert(s);
+      }}
+    />
+  );
+};
